add LocalResourceEntityHandler for uploading local HS resource data to hydroshare
This handler assumes that a local version of a HydroShare
resource (does not have to be entire resource, could just be a file in a
resource) exists locally and is stored as it is canonically in
HydroShare (i.e. {resource_id}/data/contents/). Meaning that the local
file system mirrors the resource's file system structure on HydroShare.

The handler accepts one or more files via a json object ({"files":
[str]} in the request body using the POST HTTP verb. File names passed
in the json `files` array must be in the file system context of the
passed resource. They are specified relative to `/data/contents` or
`{data}/{resource_id}` (meaning, you can include `/data/contents` if
desired.) File paths passed outside {data}/{resource_id}/ ARE NOT
ALLOWED. Additionally, file path names cannot include `~` or `..`. A 403
Forbidden is returned in the response if they are passed.

Files are zipped relative to `/data/contents` (maintains directory
structure), uploaded to HydroShare, and unzipped. Files uploaded
overwrite existing files present on HydroShare. This operation cannot be
undone.